backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "casbin>=1.41.0",
    "cryptography>=44.0.2",
    "fastapi[all]>=0.115.12",
    "openai>=1.69.0",
    "passlib[bcrypt]>=1.7.4",
    "pika>=1.3.2",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.39",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = ["watchfiles>=1.0.4"]
